% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19euxplorer-package.R
\docType{package}
\name{c19euxplorer-package}
\alias{c19euxplorer}
\alias{c19euxplorer-package}
\title{c19euxplorer: An Interactive Data Explorer App of COVID-19 in Europe}
\description{
This package features a Shiny app to explore the data of COVID-19 related variables in several countries in Europe. The variables covered are daily and cumulative cases and deaths, weekly testing and positivity rate, weekly dominant source of transmission, and daily community mobility during the pandemic.
}
\author{
\strong{Maintainer}: Dewi Lestari Amaliah \email{dama0007@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
