% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_centroids.R
\docType{data}
\name{country_centroids}
\alias{country_centroids}
\title{The centroid coordinate of countries in the world}
\format{
a tibble with 241 observations and 3 variables.
\itemize{
\item \strong{name_long}: name of the country
\item \strong{Longitude}: longitude of country's centroid
\item \strong{Latitude}: latitude of country's centroid
}
}
\source{
\href{https://worldmap.harvard.edu/data/geonode:country_centroids_az8}{Harvard World Map}
}
\usage{
country_centroids
}
\description{
A dataset contains centroid coordinate of countries in the world.
}
\keyword{datasets}
