% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_indicator.R
\name{cumulative_indicator}
\alias{cumulative_indicator}
\title{A Function to Simplify the Cumulative Cases Visualization Code}
\usage{
cumulative_indicator(df, indicator)
}
\arguments{
\item{df}{A cumulative COVID-19 tibble.}

\item{indicator}{A shiny's input_id of indicator to be visualized.}
}
\value{
A data frame containing the country, date, and the indicator.
}
\description{
This function aims to simplify the code in \code{c19euxplorer} shiny app by selecting the cumulative data
and indicator to visualized in Cases and Deaths Tab.
This function would be used in the app.R file.
}
\examples{
\dontrun{
cumulative_indicator(covid_eu_cumulative, input$indicator_select)
}

}
