% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlink.R
\name{getlink}
\alias{getlink}
\title{A Simpler Way To Create Hyperlink Text in Shiny App}
\usage{
getlink(url, label)
}
\arguments{
\item{url}{A valid URL in string}

\item{label}{A string to label the URL}
}
\description{
This function aims to simplify the typing of code to create hyperlink text in a Shiny app.
It is useful especially when there are a lot of hyperlink texts to be typed in a Shiny app's UI.
}
\examples{
getlink("https://datain360.com", "Dewi's web")

}
