% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getplot.R
\name{getplot}
\alias{getplot}
\title{Create A Line Interactive Plot of COVID-19 Variables by Date and Country}
\usage{
getplot(data_frame)
}
\arguments{
\item{data_frame}{A data frame to be visualized.}
}
\description{
This function aims to create a line plot to of COVID-19 variables by date and country.
This function could be used to visualize the data resulted from
\code{daily_indicator} or \code{cumulative_indicator} functions.
}
\examples{
\dontrun{
getplot(daily_deaths)
}

}
