% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_weekly_eu.R
\docType{data}
\name{testing_weekly_eu}
\alias{testing_weekly_eu}
\title{Weekly Europe's COVID-19 Testing and Positivity Rates}
\format{
a tibble with 954 observations and 9 variables.
\itemize{
\item \strong{country}: name of the country
\item \strong{country_code}: 2-letter ISO country code
\item \strong{year_week}: week of the year in yyyy-ww format
\item \strong{new_cases}: number of new confirmed cases
\item \strong{test_done}: number of test done
\item \strong{population}: 2019 population from the World Bank
\item \strong{testing_rate}: testing rate per 100,000 population
\item \strong{positivity_rate}: weekly test positivity (\%): 100 x Number of new confirmed cases/number of tests done per week
\item \strong{testing_data_source}: data source of testing
}
}
\source{
\href{https://www.ecdc.europa.eu/en/publications-data/covid-19-testing}{European Centre for Disease Prevention and Control}
}
\usage{
testing_weekly_eu
}
\description{
A dataset containing weekly COVID-19 testing and positivity positivity rate
of 30 countries in Europe (EU/EEA member countries and UK).
}
\keyword{datasets}
