% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_subnational_level.R
\docType{data}
\name{transmission_subnational_level}
\alias{transmission_subnational_level}
\title{Weekly COVID-19 Dominant Source of Transmission in Europe by Subnational Level}
\format{
a tibble with 11583 observations and 4 variables.
\itemize{
\item \strong{Country}: name of the country
\item \strong{Region}: subnational level, for example Province
\item \strong{Level of Transmission}: dominant source of transmission
}
\itemize{
\item Community transmission
\item Clusters of cases
\item Sporadic cases
\item Not reported i.e. the region did not reported the source of transmission
}
\itemize{
\item \strong{Week}: week from week 1 of 2020 to week 39 of 2020
}
}
\source{
\href{https://qap.ecdc.europa.eu/public/extensions/covid-19/covid-19-mobile.html#subnational-transmission-tab}{European Centre of Disease Control and Prevention}
}
\usage{
transmission_subnational_level
}
\description{
A dataset containing the dominant source of transmission of COVID-19 by
subnational level of 32 countries in Europe weekly.
}
\keyword{datasets}
