% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_selectize.R
\name{ui_selectize}
\alias{ui_selectize}
\title{Refactor User Interface selectizeInput Function}
\usage{
ui_selectize(id, choices)
}
\arguments{
\item{id}{Input id that will be used in the server side.}

\item{choices}{List of value to select.}
}
\description{
This function aim to simplify the repetitive call of shiny's selectizeInput function.
It will be called in app.R file.
}
\examples{
ui_selectize("country_select", c("Italy", "Spain", "France", "Austria", "United Kingdom"))

}
