% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_indicator.R
\name{cumulative_indicator}
\alias{cumulative_indicator}
\title{A Function to Simplify the Cumulative Cases Visualization Code}
\usage{
cumulative_indicator(
  df,
  indicator = c("Cases", "Cases per 10,000 people", "Deaths",
    "Deaths per 10,000 people")
)
}
\arguments{
\item{df}{A cumulative COVID-19 tibble.}

\item{indicator}{A shiny's input_id of indicator to be visualized.}
}
\value{
A data frame containing the country, date, and the indicator.
}
\description{
This function aims to simplify the code in \code{c19euxplorer} shiny app by selecting
country, date, and indicators in cumulative data to visualized in Cases and Deaths Tab.
This function would be used in the app.R file.
}
\details{
You can also use this function outside the app.
However, make sure to have these variables (with the exact same name)
in your data frame in order to make the function works.
\itemize{
\item cumulative_cases
\item cumulative_deaths
\item cumulative_cases_rate
\item cumulative_deaths_rate
\item country
\item date
}
}
\examples{
\dontrun{
cumulative_indicator(covid_eu_cumulative, input$indicator_select)
}

}
